refactor: add custom directive for tailwind in css, remove folder structure
