+++ client/src/App.jsx
@@ -1,7 +1,11 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Footer from "./components/Footer/Footer";
+import Footer from "./components/Footer";
 function App() {
-  return <Router></Router>;
+  return (
+    <Router>
+      <Footer />
+    </Router>
+  );
 }
 
 export default App;
